Bug Fix: Dimensional mismatch in alphago_net
A couple bug fixes were applied in networks.py.

The first bug was that values were hardcoded in the function
returning the convolutional block rather than using the values
passed to the function.

The second is that the number of filters in the convolutional block
and residual block used to be different, they actually must be the same
so now a single parameter specifies both.