Modified state2vec to make things more sensible
Modifications to state2vec and moves2vec more sensible.

state2vec now takes an additional paramater denoting whether
the rep. being generated is being done for the purposes of
evaluating the current state or generating the next possible
set of moves, this ensures that the color pane corresponds to
who just played a move.