Changing the way p and v values are calculated by the network
Changed the way that the network calculates p and v to more closely
mirror the paper in a hack-ish way.

Now a 'input batch' is treated as a single input example (so all the
state vectors generated by a call to .moves2vec() should be forward
by themselves at once).